# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Music Studio Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProject"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProject"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "204":
          description: Deleted

  # ─── Lyrics ──────────────────────────────────────────────────────────────────
  /lyrics:
    get:
      operationId: listLyrics
      tags: [lyrics]
      summary: List all lyrics
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of lyrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lyric"
    post:
      operationId: createLyric
      tags: [lyrics]
      summary: Create a lyric
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLyric"
      responses:
        "201":
          description: Created lyric
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lyric"
  /lyrics/{lyricId}:
    get:
      operationId: getLyric
      tags: [lyrics]
      summary: Get a lyric
      parameters:
        - $ref: "#/components/parameters/lyricId"
      responses:
        "200":
          description: Lyric
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lyric"
        "404":
          description: Not found
    put:
      operationId: updateLyric
      tags: [lyrics]
      summary: Update a lyric
      parameters:
        - $ref: "#/components/parameters/lyricId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLyric"
      responses:
        "200":
          description: Updated lyric
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lyric"
    delete:
      operationId: deleteLyric
      tags: [lyrics]
      summary: Delete a lyric
      parameters:
        - $ref: "#/components/parameters/lyricId"
      responses:
        "204":
          description: Deleted

  # ─── Instrumentals ───────────────────────────────────────────────────────────
  /instrumentals:
    get:
      operationId: listInstrumentals
      tags: [instrumentals]
      summary: List all instrumentals
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of instrumentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instrumental"
    post:
      operationId: createInstrumental
      tags: [instrumentals]
      summary: Create an instrumental
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstrumental"
      responses:
        "201":
          description: Created instrumental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrumental"
  /instrumentals/{instrumentalId}:
    get:
      operationId: getInstrumental
      tags: [instrumentals]
      summary: Get an instrumental
      parameters:
        - $ref: "#/components/parameters/instrumentalId"
      responses:
        "200":
          description: Instrumental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrumental"
        "404":
          description: Not found
    put:
      operationId: updateInstrumental
      tags: [instrumentals]
      summary: Update an instrumental
      parameters:
        - $ref: "#/components/parameters/instrumentalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstrumental"
      responses:
        "200":
          description: Updated instrumental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrumental"
    delete:
      operationId: deleteInstrumental
      tags: [instrumentals]
      summary: Delete an instrumental
      parameters:
        - $ref: "#/components/parameters/instrumentalId"
      responses:
        "204":
          description: Deleted

  # ─── Songs ───────────────────────────────────────────────────────────────────
...
  /releases:
    get:
      operationId: listReleases
      tags: [releases]
      summary: List all releases
      responses:
        "200":
          description: List of releases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Release"
    post:
      operationId: createRelease
      tags: [releases]
      summary: Create a release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRelease"
      responses:
        "201":
          description: Created release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
  /releases/{releaseId}:
    get:
      operationId: getRelease
      tags: [releases]
      summary: Get a release
      parameters:
        - $ref: "#/components/parameters/releaseId"
      responses:
        "200":
          description: Release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        "404":
          description: Not found
    put:
      operationId: updateRelease
      tags: [releases]
      summary: Update a release
      parameters:
        - $ref: "#/components/parameters/releaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRelease"
      responses:
        "200":
          description: Updated release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
    delete:
      operationId: deleteRelease
      tags: [releases]
      summary: Delete a release
      parameters:
        - $ref: "#/components/parameters/releaseId"
      responses:
        "204":
          description: Deleted

  # ─── Analytics ───────────────────────────────────────────────────────────────
  /analytics:
    get:
      operationId: listAnalytics
      tags: [analytics]
      summary: Get analytics entries
      parameters:
        - name: songId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Analytics list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalyticsEntry"
    post:
      operationId: createAnalyticsEntry
      tags: [analytics]
      summary: Record an analytics entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalyticsEntry"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsEntry"
...
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
    lyricId:
      name: lyricId
      in: path
      required: true
      schema:
        type: integer
    instrumentalId:
      name: instrumentalId
      in: path
      required: true
      schema:
        type: integer
...
    releaseId:
      name: releaseId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        status:
          type: string
          enum: [planning, in_progress, completed, archived]
        coverImageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, status, createdAt, updatedAt]
    CreateProject:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        status:
          type: string
          enum: [planning, in_progress, completed, archived]
          default: planning
        coverImageUrl:
          type: string
          nullable: true
      required: [name]
    UpdateProject:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        status:
          type: string
          enum: [planning, in_progress, completed, archived]
        coverImageUrl:
          type: string
          nullable: true

    # ─── Lyric ─────────────────────────────────────────────────────────────────
    Lyric:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
          nullable: true
        title:
          type: string
        content:
          type: string
        mood:
          type: string
          nullable: true
        theme:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, in_progress, complete]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, content, status, createdAt, updatedAt]
    CreateLyric:
      type: object
      properties:
        projectId:
          type: integer
          nullable: true
        title:
          type: string
        content:
          type: string
        mood:
          type: string
          nullable: true
        theme:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, in_progress, complete]
          default: draft
        notes:
          type: string
          nullable: true
      required: [title, content]
    UpdateLyric:
      type: object
      properties:
        projectId:
          type: integer
          nullable: true
        title:
          type: string
        content:
          type: string
        mood:
          type: string
          nullable: true
        theme:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, in_progress, complete]
        notes:
          type: string
          nullable: true

    # ─── Instrumental ──────────────────────────────────────────────────────────
    Instrumental:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
          nullable: true
        title:
          type: string
        producer:
          type: string
          nullable: true
        bpm:
          type: integer
          nullable: true
        musicalKey:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        mood:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        fileUrl:
          type: string
          nullable: true
        licenseType:
          type: string
          enum: [exclusive, non_exclusive, lease, free]
          nullable: true
        status:
          type: string
          enum: [available, in_use, archived]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, createdAt, updatedAt]
    CreateInstrumental:
      type: object
      properties:
        projectId:
          type: integer
          nullable: true
        title:
          type: string
        producer:
          type: string
          nullable: true
        bpm:
          type: integer
          nullable: true
        musicalKey:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        mood:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        fileUrl:
          type: string
          nullable: true
        licenseType:
          type: string
          enum: [exclusive, non_exclusive, lease, free]
          nullable: true
        status:
          type: string
          enum: [available, in_use, archived]
          default: available
        notes:
          type: string
          nullable: true
      required: [title]
    UpdateInstrumental:
      type: object
      properties:
        projectId:
          type: integer
          nullable: true
        title:
          type: string
        producer:
          type: string
          nullable: true
        bpm:
          type: integer
          nullable: true
        musicalKey:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        mood:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        fileUrl:
          type: string
          nullable: true
        licenseType:
          type: string
          enum: [exclusive, non_exclusive, lease, free]
          nullable: true
        status:
          type: string
          enum: [available, in_use, archived]
        notes:
          type: string
          nullable: true

    # ─── Song ──────────────────────────────────────────────────────────────────
...
    Release:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
          nullable: true
        title:
          type: string
        releaseType:
          type: string
          enum: [single, ep, album, mixtape]
        status:
          type: string
          enum: [planning, pre_release, released, cancelled]
        releaseDate:
          type: string
          format: date
          nullable: true
        platforms:
          type: string
          nullable: true
        distributorName:
          type: string
          nullable: true
        upc:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, releaseType, status, createdAt, updatedAt]
    CreateRelease:
      type: object
      properties:
        projectId:
          type: integer
          nullable: true
        title:
          type: string
        releaseType:
          type: string
          enum: [single, ep, album, mixtape]
          default: single
        status:
          type: string
          enum: [planning, pre_release, released, cancelled]
          default: planning
        releaseDate:
          type: string
          format: date
          nullable: true
        platforms:
          type: string
          nullable: true
        distributorName:
          type: string
          nullable: true
        upc:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [title, releaseType]
    UpdateRelease:
      type: object
      properties:
        projectId:
          type: integer
          nullable: true
        title:
          type: string
        releaseType:
          type: string
          enum: [single, ep, album, mixtape]
        status:
          type: string
          enum: [planning, pre_release, released, cancelled]
        releaseDate:
          type: string
          format: date
          nullable: true
        platforms:
          type: string
          nullable: true
        distributorName:
          type: string
          nullable: true
        upc:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # ─── Analytics ─────────────────────────────────────────────────────────────
    AnalyticsEntry:
      type: object
      properties:
        id:
          type: integer
        songId:
          type: integer
        platform:
          type: string
        streams:
          type: integer
        downloads:
          type: integer
        likes:
          type: integer
        recordedDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, songId, platform, streams, downloads, likes, recordedDate, createdAt]
    CreateAnalyticsEntry:
      type: object
      properties:
        songId:
          type: integer
        platform:
          type: string
        streams:
          type: integer
          default: 0
        downloads:
          type: integer
          default: 0
        likes:
          type: integer
          default: 0
        recordedDate:
          type: string
          format: date
      required: [songId, platform, recordedDate]
...