# lib/api-zod/src/generated/types/createRelease.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Studio Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateReleaseReleaseType } from "./createReleaseReleaseType";
import type { CreateReleaseStatus } from "./createReleaseStatus";

export interface CreateRelease {
  projectId?: number | null;
  title: string;
  releaseType: CreateReleaseReleaseType;
  status?: CreateReleaseStatus;
  releaseDate?: Date | null;
  platforms?: string | null;
  distributorName?: string | null;
  upc?: string | null;
  coverImageUrl?: string | null;
  notes?: string | null;
}