# lib/api-zod/src/generated/types/release.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Studio Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReleaseReleaseType } from "./releaseReleaseType";
import type { ReleaseStatus } from "./releaseStatus";

export interface Release {
  id: number;
  projectId?: number | null;
  title: string;
  releaseType: ReleaseReleaseType;
  status: ReleaseStatus;
  releaseDate?: Date | null;
  platforms?: string | null;
  distributorName?: string | null;
  upc?: string | null;
  coverImageUrl?: string | null;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}