# lib/api-zod/src/generated/types/updateInstrumental.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Studio Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateInstrumentalLicenseType } from "./updateInstrumentalLicenseType";
import type { UpdateInstrumentalStatus } from "./updateInstrumentalStatus";

export interface UpdateInstrumental {
  projectId?: number | null;
  title?: string;
  producer?: string | null;
  bpm?: number | null;
  musicalKey?: string | null;
  genre?: string | null;
  mood?: string | null;
  durationSeconds?: number | null;
  fileUrl?: string | null;
  licenseType?: UpdateInstrumentalLicenseType;
  status?: UpdateInstrumentalStatus;
  notes?: string | null;
}